backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
motor
pymongo
celery[redis]
redis
requests